+++ MCT/Controllers/TournamentsController.cs
@@ -18,42 +18,31 @@ namespace MCT.Controllers
             _context = context;
         }
 
-        // GET: Tournaments
         public async Task<IActionResult> Index()
         {
             var mctContext = _context.Tournaments.Include(t => t.StatusNavigation);
             return View(await mctContext.ToListAsync());
         }
 
-        // GET: Tournaments/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var tournament = await _context.Tournaments
                 .Include(t => t.StatusNavigation)
                 .FirstOrDefaultAsync(m => m.TournamentId == id);
-            if (tournament == null)
-            {
-                return NotFound();
-            }
+
+            if (tournament == null) return NotFound();
 
             return View(tournament);
         }
 
-        // GET: Tournaments/Create
         public IActionResult Create()
         {
             ViewData["Status"] = new SelectList(_context.TournamentStatuses, "StatusName", "StatusName");
             return View();
         }
 
-        // POST: Tournaments/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("TournamentId,Description,Location,StartDate,EndDate,Price,Status")] Tournament tournament)
@@ -68,34 +57,22 @@ namespace MCT.Controllers
             return View(tournament);
         }
 
-        // GET: Tournaments/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var tournament = await _context.Tournaments.FindAsync(id);
-            if (tournament == null)
-            {
-                return NotFound();
-            }
+            if (tournament == null) return NotFound();
+
             ViewData["Status"] = new SelectList(_context.TournamentStatuses, "StatusName", "StatusName", tournament.Status);
             return View(tournament);
         }
 
-        // POST: Tournaments/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("TournamentId,Description,Location,StartDate,EndDate,Price,Status")] Tournament tournament)
         {
-            if (id != tournament.TournamentId)
-            {
-                return NotFound();
-            }
+            if (id != tournament.TournamentId) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -106,14 +83,8 @@ namespace MCT.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!TournamentExists(tournament.TournamentId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!TournamentExists(tournament.TournamentId)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -121,35 +92,25 @@ namespace MCT.Controllers
             return View(tournament);
         }
 
-        // GET: Tournaments/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var tournament = await _context.Tournaments
                 .Include(t => t.StatusNavigation)
                 .FirstOrDefaultAsync(m => m.TournamentId == id);
-            if (tournament == null)
-            {
-                return NotFound();
-            }
+
+            if (tournament == null) return NotFound();
 
             return View(tournament);
         }
 
-        // POST: Tournaments/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var tournament = await _context.Tournaments.FindAsync(id);
-            if (tournament != null)
-            {
-                _context.Tournaments.Remove(tournament);
-            }
+            if (tournament != null) _context.Tournaments.Remove(tournament);
 
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
